+++ src/middleware.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express";
+import { HttpError } from "http-errors";
 import { Logger } from "winston";
 import { z } from "zod";
 import { FlatObject, IOSchema } from "./common-helpers";
@@ -31,3 +32,18 @@ export const createMiddleware = <
 >(
   definition: MiddlewareDefinition<IN, OPT, OUT>
 ) => definition;
+
+export type ExpressMiddleware<R extends Request, S extends Response> = (
+  request: R,
+  response: S,
+  next: (error?: any) => void
+) => void | Promise<void>;
+
+export interface ExpressMiddlewareFeatures<
+  R extends Request,
+  S extends Response,
+  OUT extends FlatObject
+> {
+  provider?: (request: R, response: S) => OUT | Promise<OUT>;
+  transformer?: (err: Error) => HttpError | Error;
+}