+++ tests/system/system.spec.ts
@@ -1,3 +1,4 @@
+import cors from "cors";
 import http from "http";
 import fetch from "node-fetch";
 import {
@@ -17,6 +18,24 @@ describe("App", () => {
   beforeAll(() => {
     const routing = {
       v1: {
+        corsed: new EndpointsFactory(defaultResultHandler)
+          .use(
+            cors({
+              credentials: true,
+              exposedHeaders: ["Content-Range", "X-Content-Range"],
+            }),
+            {
+              provider: () => ({ corsDone: true }),
+            }
+          )
+          .build({
+            method: "get",
+            input: z.object({}),
+            output: z.object({ corsDone: z.boolean() }),
+            handler: async ({ options }) => ({
+              corsDone: options.corsDone,
+            }),
+          }),
         faulty: new EndpointsFactory(
           createResultHandler({
             getPositiveResponse: () => createApiResponse(z.object({})),
@@ -83,7 +102,7 @@ describe("App", () => {
           listen: 8055,
           compression: { threshold: 1 },
         },
-        cors: true,
+        cors: false,
         startupLogo: true,
         logger: {
           level: "silent",
@@ -191,6 +210,24 @@ describe("App", () => {
       const json = await response.json();
       expect("status" in json).toBeTruthy();
       expect(json.status).toBe("success");
+    });
+
+    test("Should execute native express middleware", async () => {
+      const response = await fetch("http://127.0.0.1:8055/v1/corsed", {
+        method: "GET",
+      });
+      expect(response.status).toBe(200);
+      const json = await response.json();
+      expect(json).toEqual({
+        status: "success",
+        data: { corsDone: true },
+      });
+      expect(response.headers.get("Access-Control-Allow-Credentials")).toBe(
+        "true"
+      );
+      expect(response.headers.get("Access-Control-Expose-Headers")).toBe(
+        "Content-Range,X-Content-Range"
+      );
     });
   });